+++ app/src/main/java/me/devsaki/hentoid/activities/BaseWebActivity.java
@@ -27,6 +27,7 @@ import me.devsaki.hentoid.enums.StatusContent;
 import me.devsaki.hentoid.services.DownloadService;
 import me.devsaki.hentoid.util.Consts;
 import me.devsaki.hentoid.util.ConstsImport;
+import me.devsaki.hentoid.util.FileHelper;
 import me.devsaki.hentoid.util.Helper;
 import me.devsaki.hentoid.util.LogHelper;
 import me.devsaki.hentoid.views.ObservableWebView;
@@ -251,7 +252,7 @@ public class BaseWebActivity extends BaseActivity {
             currentContent = db.selectContentById(currentContent.getId());
             if (StatusContent.DOWNLOADED == currentContent.getStatus()
                     || StatusContent.ERROR == currentContent.getStatus()) {
-                Helper.openContent(this, currentContent);
+                FileHelper.openContent(this, currentContent);
             } else {
                 hideFab(fabRead);
             }
