+++ app/src/main/java/me/devsaki/hentoid/listener/ItemClickListener.java
@@ -6,6 +6,7 @@ import android.view.View.OnClickListener;
 import android.view.View.OnLongClickListener;
 
 import me.devsaki.hentoid.database.domains.Content;
+import me.devsaki.hentoid.util.FileHelper;
 import me.devsaki.hentoid.util.Helper;
 import me.devsaki.hentoid.util.LogHelper;
 
@@ -48,7 +49,7 @@ public class ItemClickListener implements OnClickListener, OnLongClickListener {
     public void onClick(View v) {
         if (!selected) {
             Helper.toast(context, "Opening: " + content.getTitle());
-            Helper.openContent(context, content);
+            FileHelper.openContent(context, content);
         }
     }
